TIP-453: introduce the Database/AbstractReader
