+++ src/Pim/Component/Connector/Reader/Database/AbstractReader.php
@@ -1,0 +1,70 @@
+<?php
+
+namespace Pim\Component\Connector\Reader\Database;
+
+use Akeneo\Component\Batch\Item\AbstractConfigurableStepElement;
+use Akeneo\Component\Batch\Item\ItemReaderInterface;
+use Akeneo\Component\Batch\Model\StepExecution;
+use Akeneo\Component\Batch\Step\StepExecutionAwareInterface;
+use Doctrine\Common\Persistence\ObjectRepository;
+
+/**
+ * Abstract reader
+ *
+ * @author    Julien Janvier <jjanvier@akeneo.com>
+ * @copyright 2016 Akeneo SAS (http://www.akeneo.com)
+ * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
+ */
+abstract class AbstractReader extends AbstractConfigurableStepElement implements
+    ItemReaderInterface,
+    StepExecutionAwareInterface
+{
+    /** @var bool Checks if all objects are sent to the processor */
+    protected $isExecuted = false;
+
+    /** @var StepExecution */
+    protected $stepExecution;
+
+    /** @var \ArrayIterator */
+    protected $results;
+
+    /**
+     * {@inheritdoc}
+     */
+    public function read()
+    {
+        if (!$this->isExecuted) {
+            $this->isExecuted = true;
+
+            $this->results = $this->getResults();
+        }
+
+        if (null !== $result = $this->results->current()) {
+            $this->results->next();
+            $this->stepExecution->incrementSummaryInfo('read');
+        }
+
+        return $result;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function setStepExecution(StepExecution $stepExecution)
+    {
+        $this->stepExecution = $stepExecution;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function initialize()
+    {
+        $this->isExecuted = false;
+    }
+
+    /**
+     * @return \ArrayIterator
+     */
+    abstract protected function getResults();
+}
